move s2k Config into a separate file
